64 bits fixes integration
git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_3_0_0_dev@353 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/include/tulip/TreeTest.h
@@ -43,7 +43,7 @@ private:
   void destroy(SuperGraph *);
   TreeTest();
   static TreeTest * instance;
-  stdext::hash_map<unsigned int,bool> resultsBuffer;
+  stdext::hash_map<unsigned long,bool> resultsBuffer;
 };
 /*@}*/
 #endif
